requirements.txt
streamlit 
psycopg2-binary 
langchain-google-genai
 pgvector
  sentence-transformers 
  sqlalchemy 
  pandas
   plotly
   dotenv